fix (refactor humanSvg component):
1) move tooltip methods to separate utils
2) add new utils to: HumanSvgComponent.js and Search.js
3) don't get 10 elements by ids from svg: instead re-use pathTriggers
4) remove reduce/maps over already existing arrays
5) have only single loop over pathtriggers
6) remove switch/case over tissues (10 cases): reuse pathtriggers
