+++ hubmap/frontend/src/commons/animation-utils.js
